+++ yash-env/CHANGELOG.md
@@ -28,6 +28,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `stack::Frame` is now `non_exhaustive`.
 - `system::virtual::FileSystem::get` now fails with `EACCES` when search
   permission is denied for any directory component of the path.
+- The following methods of `system::virtual::Process` now operate on
+  `signal::Number` instead of `trap::Signal`:
+    - `signal_handling`
+    - `set_signal_handling`
 - The type parameter constraint for `subshell::Subshell` is now
   `F: for<'a> FnOnce(&'a mut Env, Option<JobControl>) -> Pin<Box<dyn Future<Output = ()> + 'a>> + 'static`.
   The `Output` type of the returned future has been changed from